requirements.txt
pyserial
asyncio
